+++ Source/Sandbox/SampleLibrary/UnionFind.cs
@@ -1,4 +1,5 @@
-﻿using AtCoder;
+﻿using System.Diagnostics.CodeAnalysis;
+using AtCoder;
 
 namespace SampleLibrary
 {
@@ -7,6 +8,16 @@ namespace SampleLibrary
         public UnionFind(int n) : base(n) { Foo(); }
 
         void Foo() => Bar();
+        public bool Try([NotNullWhen(true)] out string? text)
+        {
+            if (this.Size(0) == 1)
+            {
+                text = "Single";
+                return true;
+            }
+            text = null;
+            return false;
+        }
         partial void Bar();
     }
     [SourceExpander.NotEmbeddingSource]