+++ Source/SourceExpander.Embedder/Diagnostics/DiagnosticsResources.Designer.cs
@@ -19,7 +19,7 @@ namespace SourceExpander.Diagnostics {
     // または Visual Studio のようなツールを使用して自動生成されました。
     // メンバーを追加または削除するには、.ResX ファイルを編集して、/str オプションと共に
     // ResGen を実行し直すか、または VS プロジェクトをビルドし直します。
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "16.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     internal class DiagnosticsResources {
@@ -165,24 +165,6 @@ namespace SourceExpander.Diagnostics {
         internal static string EMBED0006_Title {
             get {
                 return ResourceManager.GetString("EMBED0006_Title", resourceCulture);
-            }
-        }
-        
-        /// <summary>
-        ///   Nullable option is unsupported に類似しているローカライズされた文字列を検索します。
-        /// </summary>
-        internal static string EMBED0007_Title {
-            get {
-                return ResourceManager.GetString("EMBED0007_Title", resourceCulture);
-            }
-        }
-        
-        /// <summary>
-        ///   Nullable directive is unsupported に類似しているローカライズされた文字列を検索します。
-        /// </summary>
-        internal static string EMBED0008_Title {
-            get {
-                return ResourceManager.GetString("EMBED0008_Title", resourceCulture);
             }
         }